+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 * Serializing to .tsv via serde
 
 * Rename `gams range` to `gams rg`
-* Rename `gams fsw` to `gams swstat`
+* Rename `gams sliding` to `gams wave`
+* Rename `gams fsw` to `gams sw`
 * Add --seq to `gams locate`
 * Add --count to `gams locate`
 * Add --parallel to `gams sliding` and `gams fsw`